Fix epic callpeak command, and other changes
The epic command included a "tail -f" in order to log to both a file and
stdout, but I forgot that "tail -f" doesn't know when the log file is
"done" and therefore never exits. In addition, I've added versions to
the IDR and CCF rules, and converted a bunch of log files to the "log:"
property instead of just making them another output file.