Fix an issue with idr input file processing
The idr script can only use the "score" column with an input file type
of bed, so even though the input from Epic is in narrowPeak format, we
tell idr that it's bed format (a subset of the columns in narrowPeak) in
order to use the score column. For consistency, all IDR output files for
both MACS2 and Epic now use the bed format.