Use only BiocParallel for parallelism
This ensures that all parallel operations use the same backend.

This commit also add a better parallelization scheme for
featureCounts, using BiocParallel. The nthreads option to
featureCounts seems to have no effect on my system. Perhaps the file
I/O is the bottleneck or something. In any case, this now uses
BiocParallel to run featureCounts in a separate process for each file,
and then merge the results.