Improve parametrization of RNA-seq file names
From the SummarizedExperiment stage on, we now always use
hg38.analysisSet as the genome name, even for HISAT2, where the
alignment was done against grch38_snp_tran, since the quantification
is done with the same reference as all the other quantification
methods. This simplifies the downstream file name handling, since we
don't have to handle the special case of HISAT2 using a different
genome name from all the other methods.